Add files via upload
Alternative formats for MiniGui.chm

MiniGui.exe, is a self-contained executable Help file. This, for me, is the most convenient format.

MiniGui.html is all of the Help Text in a single Html file.

MiniGui.dbf and fpt - Might be useful, no idea.

MiniGui.mdb - Microsoft Access

MiniGui.pdf

Note that the converted Help file Minigui.exe includes only text, and not images.

These Help files are current as of 01/19/2022.